+++ src/pages/NewsPage.jsx
@@ -1,15 +1,9 @@
-import React, {useEffect, useState} from 'react';
-import {useDispatch, useSelector} from 'react-redux';
+import React, {useState, useEffect} from 'react';
 import styled from 'styled-components';
-import {Swiper, SwiperSlide} from 'swiper/react';
-import 'swiper/swiper-bundle.css';
-import {useLocation, useNavigate} from 'react-router-dom';
+import {useNavigate, useLocation} from 'react-router-dom';
 import CategoryButton from '../components/CategoryButton';
-import LoadingSpinner from '../components/LoadingSpinner';
-import ModalSection from '../components/ModalSection';
 import NewsPagination from '../components/News/NewsPagination';
-import {NewsApi, NewsMonthApi} from '../lib/Apiservice';
-import {setMonthTitle} from '../redux/monthFieldSlice';
+import {NewsApi} from '../lib/Apiservice';
 
 const NewsMain = styled.section`
   height: calc(100vh - 4.5rem - 140.78px);
@@ -22,30 +16,16 @@ const H1 = styled.h1`
   text-align: center;
 `;
 
-const SwiperContainer = styled.div`
-  width: 100%;
-  margin: 3rem 0;
-`;
-
-const Image = styled.img`
-  width: 100%;
-  object-fit: cover;
-  border-radius: 1rem;
-`;
-
 const ButtonWrapper = styled.div`
   margin: 0 5%;
 `;
 
-function NewsPage() {
+export default function NewsPage() {
   const navigate = useNavigate();
-  const dispatch = useDispatch();
   const location = useLocation();
-  const monthFieldTitle = useSelector(state => state.monthTitle.value);
-  const [loading, setLoading] = useState(true);
-  const [showedMonthField, setShowedMonthField] = useState({});
-  const [selectCategory, setSelectCategory] = useState('월간필드');
   const categoryArr = ['월간필드', '취업/진로', 'FIELD', '공모전'];
+  const [selectCategory, setSelectCategory] = useState('월간필드');
+  const [loading, setLoading] = useState(true);
   const [newsData, setNewsData] = useState([]);
   const imageUrl = `${import.meta.env.VITE_API_URL}/api/files/damzbyg116zhar4/`;
   const handleButtonClick = item => {
@@ -53,49 +33,14 @@ function NewsPage() {
     navigate(`/news?category=${item}`);
   };
 
-  const fetchNewsData = async category => {
-    let response = JSON.parse(localStorage.getItem(category));
-    if (!response) {
-      response = await NewsApi(category);
-      localStorage.setItem(category, JSON.stringify(response));
-    }
-    return response;
-  };
-
-  const updateMonthFieldTitle = (response, category, title) => {
-    if (category === '월간필드' && title === '') {
-      const titleArr = response.map(item => item.title);
-      dispatch(setMonthTitle(titleArr[titleArr.length - 1]));
-    }
-  };
-
   const getDataNews = async () => {
     try {
-      const response = await fetchNewsData(selectCategory);
-      updateMonthFieldTitle(response, selectCategory, monthFieldTitle);
+      const response = await NewsApi(selectCategory);
       setNewsData(response);
-      setLoading(false);
-    } catch (err) {
-      console.log(err);
-    }
-  };
-
-  const getDataMonthField = async () => {
-    setLoading(true);
-    const monthLocalData = JSON.parse(localStorage.getItem('monthFieldTitle'));
-    let response;
-    try {
-      if (monthLocalData && monthFieldTitle === monthLocalData[0].title) {
-        response = monthLocalData;
-      } else {
-        response = await NewsMonthApi(monthFieldTitle);
-        localStorage.setItem('monthFieldTitle', JSON.stringify(response));
-      }
-      setShowedMonthField(response[0]);
     } catch (err) {
       console.log(err);
     } finally {
-      setLoading(false); // 데이터 로딩 완료
+      setLoading(false);
     }
   };
 
@@ -108,45 +53,6 @@ function NewsPage() {
     setSelectCategory(urlCategory);
   }, [location.search]);
 
-  useEffect(() => {
-    if (selectCategory === '월간필드' && monthFieldTitle !== '') {
-      getDataMonthField(monthFieldTitle);
-    }
-  }, [monthFieldTitle, selectCategory]);
-
-  let content;
-
-  if (loading) {
-    content = <LoadingSpinner />;
-  } else if (selectCategory === '월간필드' && showedMonthField?.photo) {
-    content = (
-      <>
-        <ModalSection
-          timeDatalst={newsData.map(item => item.title)}
-          title={monthFieldTitle}
-          color='yellow'
-          $margin='3rem 0'
-        />
-        <SwiperContainer>
-          <Swiper
-            slidesPerView={1.2}
-            spaceBetween={20}
-            initialSlide={0}
-            centeredSlides='true'
-            key={monthFieldTitle}
-          >
-            {showedMonthField.photo.map(item => (
-              <SwiperSlide key={item}>
-                <Image src={`${imageUrl}${showedMonthField.id}/${item}`} alt={showedMonthField} />
-              </SwiperSlide>
-            ))}
-          </Swiper>
-        </SwiperContainer>
-      </>
-    );
-  } else if (selectCategory !== '월간필드') {
-    content = <NewsPagination newsData={newsData} category={selectCategory} />;
-  }
   return (
     <NewsMain>
       <H1>NEWS</H1>
@@ -160,9 +66,7 @@ function NewsPage() {
           />
         ))}
       </ButtonWrapper>
-      {content}
+      <NewsPagination newsData={newsData} category={selectCategory} />;
     </NewsMain>
   );
 }
-
-export default NewsPage;
